refactor: separate docs shell asset generators
